main/java/com/ecse428/jad/amv/MyActivity.java
package com.ecse428.jad.amv;

import android.support.v7.app.ActionBarActivity;
import android.os.Bundle;
import android.view.Menu;
import android.view.MenuItem;


public class MyActivity extends ActionBarActivity {

    @Override
    protected void onCreate(Bundle savedInstanceState) {
        super.onCreate(savedInstanceState);
        setContentView(R.layout.activity_my);
    }


    @Override
    public boolean onCreateOptionsMenu(Menu menu) {
        // Inflate the menu; this adds items to the action bar if it is present.
        getMenuInflater().inflate(R.menu.menu_my, menu);
        return true;
    }

    @Override
    public boolean onOptionsItemSelected(MenuItem item) {
        // Handle action bar item clicks here. The action bar will
        // automatically handle clicks on the Home/Up button, so long
        // as you specify a parent activity in AndroidManifest.xml.
        int id = item.getItemId();

        //noinspection SimplifiableIfStatement
        if (id == R.id.action_settings) {
            return true;
        }

        return super.onOptionsItemSelected(item);
    }

    public Vector vectorAdd(Vector v1, Vector v2){

        return new Vector(v1.x + v2.x, v1.y + v2.y);
    }

    public Double scalarMult(Vector v1, Vector v2){

        return (v1.x*v2.x + v1.y*v2.y);
    }

    public Double vectorMult(Vector v1, Vector v2){

        return (v1.x*v2.y - v1.y*v2.x);
    }

}
